Implement simple command-line serving via httpd
Often times people want to quickly have a directory index of some folder
to either view directly or share on the local network. For instance, in
Python, `python3 -m http.server` implements a simple server
accomplishing exactly this.

`httpd` has the needed functionality to do this, but did not define a
command-line entrypoint. Users who wanted to quickly serve files with
Erlang needed hacks like shown here:
https://gist.github.com/willurd/5720255#erlang. This commit adds support
for the following forms of hosting a simple directory listing with
support for `index.html` and basic mimetypes, with the following
commands:

    erl -s httpd
    erl -s httpd serve
    erl -s httpd serve path/to/dir

Instead of `-s`, `-run` is also supported. The original suggestion
cropped up here:
https://erlangforums.com/t/httpc-httpd-improvements/2622/8