Allow running tests more silently
When running tests at present via `make test`, the developer is met with
lots of messages from both the testing script and logger messages
emitted through libraries called by the testcases themselves. This
clutters the output and makes it hard to distinguish actual test suite
failures from regular progress and other logging messages.

This commit introduces support for interpreting the `-s` / `--silent` /
`--quiet` option from `make` and utilizing it to run tests in a more
quiet manner. Specifically, this means that log output is fully captured
by `cth_log_redirect` (instead of logging to both stdout and the CT
log), and the banners and disclaimers from the testing script are
reduced to a minimum (under the assumption that the person running the
tests is familar with them).

To accomplish this, the `cth_log_redirect` handler is augmented with the
`mode` option, which instructs it how to work. The previous behaviour -
logging to both stdout and the Common Test logs - is kept by default for
now.

As a small improvement that came up whilst implementing this change,
`logger_std_h` is updated to skip logging filesync errors when
instructeed to log to `/dev/null`.

In `inets`, this reduces the output emitted by the tests from 2175 to
803 lines.