gh: Add annotations to failed tests
Previously, when running the tests outside of a pull request in a fork,
finding the failed test result would involve downloading the test logs
and opening them locally.

With this commit, the JUnit report generated by `cth_surefire` is
used to add annotations to the files in GitHub, visible from the commit
checks view or in the PR diff.